+++ torchao/ops.py
@@ -197,19 +197,17 @@ def _(
 
 def to_sparse_semi_structured_cutlass_sm9x_f8(
     weight: Tensor,
-    backend: str = "legacy",
 ) -> (Tensor, Tensor):
-    if backend == "legacy":
-        return torch.ops.torchao.to_sparse_semi_structured_cutlass_sm9x_f8.default(
-            weight
-        )
-    if backend == "cutedsl":
-        from torchao.quantization.quantize_.workflows.float8.cutedsl_sparse_2x4 import (
-            to_sparse_semi_structured_cutedsl,
-        )
+    """
+    CUTLASS-based conversion from sparsified input tensor to corresponding compressed tensor, along with corresponding metadata tensor.
+    Args:
+        weight: input tensor, in row-major layout.
+    Returns:
+        weight_compressed: compressed weight tensor, with sparsity eliminated, in row-major layout.
+        weight_meta: metadata tensor, describing the sparsity structure of the input tensor, also in row-major layout.
+    """
 
-        return to_sparse_semi_structured_cutedsl(weight)
-    raise ValueError(f"Unsupported sparse conversion backend: {backend}")
+    return torch.ops.torchao.to_sparse_semi_structured_cutlass_sm9x_f8.default(weight)
 
 
 @register_custom_op("torchao::to_sparse_semi_structured_cutlass_sm9x_f8")
